Fix: Class "RecurringPolicy" has 25 methods, which is greater than 20 authorized. Split it into smaller classes.
